requirements.txt
anthropic>=0.51.0
arxiv>=2.2.0
mcp>=1.7.1
pypdf2>=3.0.1
python-dotenv>=1.1.0
uv
groq